Asteroids spawn and split appropriately(almost)

+++ Assets/Scripts/Managers/AsteroidManager.cs
@@ -4,36 +4,44 @@ using ExtensionMethods;
 public class AsteroidManager : MonoBehaviour
 {
     private const int MAX_ASTEROID_SIZE = 3;
-    private const int NUM_SPLIT_ASTEROIDS = 2;
+    private const int NUM_SPLIT_ASTEROIDS = 3;
 
     public int Level;
 
     private ObjectPool<Asteroid> asteroidPool;
 
-    private AsteroidCollisionEvent asteroidCollisionEvent = new AsteroidCollisionEvent();
-
     void Awake()
     {
         asteroidPool = new ObjectPool<Asteroid>();
-        int poolSize = Mathf.CeilToInt(Mathf.Pow(NUM_SPLIT_ASTEROIDS, MAX_ASTEROID_SIZE));
-        asteroidPool.Initialize(poolSize, transform);
-
-        asteroidCollisionEvent.AddListener(OnAsteroidCollisionEvent);
     }
 
     private void Start()
     {
+        Initialize();
+    }
+
+    private void Initialize()
+    {
+        int poolSize = Level * Mathf.CeilToInt(Mathf.Pow(NUM_SPLIT_ASTEROIDS, MAX_ASTEROID_SIZE));
+        asteroidPool.Initialize(poolSize, transform);
+
         SpawnAsteroids(Level, MAX_ASTEROID_SIZE, Vector2.zero);
     }
 
     private void OnAsteroidCollisionEvent(Asteroid asteroid, string colliderTag)
     {
-        asteroidPool.Kill(asteroid);
+        if (colliderTag == Tags.ASTEROID_TAG)
+            return;
 
         if (asteroid.Size > 0)
         {
-            SpawnAsteroids(NUM_SPLIT_ASTEROIDS, asteroid.Size - 1, asteroid.RB.position);
+            int newSize = asteroid.Size - 1;
+            SpawnAsteroids(NUM_SPLIT_ASTEROIDS, newSize, asteroid.Position);
         }
+
+        asteroid.AsteroidCollisionEvent.RemoveListener(OnAsteroidCollisionEvent);
+
+        asteroidPool.Kill(asteroid);
     }
 
     private void SpawnAsteroids(int amount, int size, Vector2 location)
@@ -43,13 +51,23 @@ public class AsteroidManager : MonoBehaviour
         for (int i = 0; i < amount; i++)
         {
             var asteroid = asteroidPool.Spawn();
-            asteroid.Size = size;
-            asteroid.RB.position = positionArray[i];
+
+            if (asteroid == null)
+                continue;
 
             var asteroidSpeed = this.GetRandomInRange(asteroid.MinSpeed, asteroid.MaxSpeed);
-            var asteroidDirection = this.GetDirectionInRangeRelatedToPosition(amount, location, asteroid.RB.position);
+            var asteroidDirection = this.GetDirectionInRangeRelatedToPosition(amount, location, asteroid.Position);
+            var asteroidVelocity = asteroidSpeed * asteroidDirection;
 
-            asteroid.RB.velocity = asteroidSpeed * asteroidDirection;
+            asteroid.Initialize(size, positionArray[i], asteroidVelocity);
+
+            asteroid.AsteroidCollisionEvent.AddListener(OnAsteroidCollisionEvent);
         }
+    }
+
+    [ContextMenu("Restart Level")]
+    public void RestartLevel()
+    {
+        Initialize();
     }
 }
